requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
codecov==2.0.15
coverage==4.5.4
cryptography==2.7
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.1.0
django-countries==5.4
django-filter==2.2.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
idna==2.8
importlib-metadata==0.19
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
more-itertools==7.2.0
mysqlclient==1.4.4
ndg-httpsclient==0.5.1
oauthlib==3.1.0
packaging==19.1
Pillow==6.1.0
pluggy==0.12.0
py==1.8.0
pyasn1==0.4.6
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pylint-plugin-utils==0.5
pyOpenSSL==19.0.0
pyparsing==2.4.2
pytest==5.0.1
pytest-cov==2.7.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.5.2
